# OnlineCinema/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Фильмотека",
        "title": "Фильмотека API",
        "contact": {},
        "version": "1.0"
    },
    "host": ":8080",
    "paths": {
...
        "/CreateFilm": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "добавляет информацию о Фильме",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Добавить фильм",
                "operationId": "CreateFilm",
                "parameters": [
                    {
                        "description": "Информация о фильме",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Film"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/CreateUser": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Создает нового пользователя приложения",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Создать пользователя",
                "operationId": "CreateUser",
                "parameters": [
                    {
                        "description": "Информация о пользователе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/DeleteFilm": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "удаляет информацию о Фильме",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "удалить фильм",
                "operationId": "DeleteFilm",
                "parameters": [
                    {
                        "description": "Информация о фильме",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Film"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/DeleteUser": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Удаляет пользователя приложения",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Удалить пользователя",
                "operationId": "DeleteUser",
                "parameters": [
                    {
                        "description": "Идентификатор пользователя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.DeleteUserStruct"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/GetAllUsers": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Возвращает всех пользователей приложения",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "получить пользователей",
                "operationId": "GetAllUsers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        },
        "/GetFilms": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Получает информацию о фильмах",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Показать фильмы",
                "operationId": "GetFilms",
                "parameters": [
                    {
                        "description": "условия выборки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.GetFilmsParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.OutFilm"
                        }
                    }
                }
            }
        },
...
        "/UpdateFilm": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "обновляет информацию о Фильме",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Обновить фильм",
                "operationId": "UpdateFilm",
                "parameters": [
                    {
                        "description": "Информация о фильме",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Film"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "storage.Actor": {
            "type": "object",
            "properties": {
                "DeleteFilms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "actorname": {
                    "type": "string"
                },
                "addfilms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "birthdate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
        "storage.Film": {
            "type": "object",
            "properties": {
                "addActors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "deleteActors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "filmname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "string"
                },
                "releasedate": {
                    "type": "string"
                }
            }
        },
        "storage.GetFilmsParam": {
            "type": "object",
            "properties": {
                "actorname": {
                    "type": "string"
                },
                "filmname": {
                    "type": "string"
                },
                "sortname": {
                    "type": "string"
                },
                "sorttype": {
                    "type": "string"
                }
            }
        },
...
        "storage.OutFilm": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Actor"
                    }
                },
                "description": {
                    "type": "string"
                },
                "filmname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "string"
                },
                "releasedate": {
                    "type": "string"
                }
            }
        },
        "users.DeleteUserStruct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "users.User": {
            "description": "User информация о пользователе",
            "type": "object",
            "properties": {
                "id": {
                    "description": "идентификатор пользователя",
                    "type": "integer"
                },
                "password": {
                    "description": "пароль пользователя",
                    "type": "string"
                },
                "userIsAdmin": {
                    "description": "признак того, что пользователь является администратором",
                    "type": "boolean"
                },
                "username": {
                    "description": "имя пользователя",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}